requirements.txt
# Core Framework
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==21.2.0

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Visualization
plotly==5.24.1

# Graph Analysis (for train movement module)
networkx==3.3

# Configuration
python-dotenv==1.0.1

# Date/Time Utilities
python-dateutil==2.8.2